Refactors clipboard and editor functionality
Moves clipboard and editor-related functions into separate utility modules.

This improves code organization and promotes reusability by centralizing
these common operations. It also allows for more focused unit testing of
these utilities in the future.
